(mail-default-headers): Added an "X-Mailer" line (mail-summary-lines): reduced to eight
(mail-parse-list, mail-parse-group, mail-parse-atom): the second and
third functions are new, split off from mail-parse-list. It's now
possible to read "groups" of expressions---things not separated by
commas. A new optional arg to mail-parse-list controls this

(mail-get-header): option to read a list of groups, not a list of
comma-separated expressions

(mail-insert-list): option to suppress comma-separators