(paragraph-regexp): renamed as paragraph-separate, this is now a regexp that matches the beginning of a line that separates two paragraphs
(page-regexp): renamed as page-start

(forward-paragraph, backward-paragraph): totally rewritten, these work
properly now. As with all other forward- and backward- functions,
forward-paragraph find the character following the end of the
paragraph, backward-paragraph finds the character at the start of the
paragraph

(mark-paragraph): renamed as paragraph-edges, this function now finds
the beginning and end of the current COUNT paragraphs. Marks them as a
block when called interactively.

(transpose-paragraphs): does what you'd expect, bound to `Ctrl-x
Ctrl-Meta-t'.

(save-cursor): now uses a mark to store the position