renamed from fill-mode.jl, this file has been almost completely rewritten
(fill-break-re): regexp defining where it's permissable to break lines
(by default: space, tab, newline or form-feed)

(fill-mode-active): what used to be fill-mode-p

(fill-prefix, fill-prefix-width): variable defining the fill prefix of
the buffer. A string or a function, see its documentation for details.

(set-fill-prefix): new command (bound to `Ctrl-x .') that saves from
the start of the line to the cursor as the fill prefix for the buffer

(fill-area): new function to fill an arbitrary region of text. Handles
a prefix at the beginning of each line

(fill-paragraph): now calls fill-area

(fill-check-line): rewritten (this is for auto-fill mode)

(center-line, center-paragraph): moved from text-mode.jl, untested